Use double for weight.

+++ src/engine/gameobjects/item.cpp
@@ -11,7 +11,7 @@ Item::Item(Realm *realm, GameObjectType objectType, uint id, Options options) :
     super(realm, objectType, id, objectType == GameObjectType::Player ? options :
                                  (Options) (options | AutomaticNameForms)),
     m_position(0, 0, 0),
-    m_weight(0),
+    m_weight(0.0),
     m_cost(0.0),
     m_flags(ItemFlags::NoFlags) {
 }
@@ -28,7 +28,7 @@ void Item::setPosition(const Point3D &position) {
     }
 }
 
-void Item::setWeight(int weight) {
+void Item::setWeight(double weight) {
 
     if (m_weight != weight) {
         m_weight = weight;
